refactor: split `util.ts` file into `is-of-type.ts` and `is-archived.ts`
