+++ apps/hash-frontend/src/pages/[shortname]/shared/archive-menu-item.tsx
@@ -11,8 +11,8 @@ import {
 import { archiveEntityTypeMutation } from "../../../graphql/queries/ontology/entity-type.queries";
 import { useFetchEntityTypes } from "../../../shared/entity-types-context/hooks";
 import { BoxArchiveIcon } from "../../../shared/icons/box-archive-icon";
+import { isEntityPageEntity } from "../../../shared/is-of-type";
 import { MenuItem } from "../../../shared/ui/menu-item";
-import { isEntityPageEntity } from "../../../shared/util";
 import {
   isItemEntityType,
   useContextBarActionsContext,
